added support for transforming examples with HelloWorld template values
